Moved tests to new files, no code change
